rename {SpectralFunctions => spectral_functions}.hpp
