+++ include/spectral_functions.hpp
